refactor(cli): extract context parser args
